Fix an issue where garbage collection bloats DB
There is an issue I keep hitting where the `cron` value in `wp_options`
get's so big that the site no longer operates.

I believe this could be due to an issue mentioned at
https://codex.wordpress.org/Function_Reference/wp_schedule_event

'For some reason there seems to be a problem on some systems where the
hook must not contain underscores or uppercase characters.'

The issue causes the hook to be repeatedly added to the serialised
array, which then means WordPress times out trying to communicate with
the database.